Refactor: Extract pagination logic into a custom hook
